backend/phishing_detector/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.13
djangorestframework==3.15.2
idna==3.7
joblib==1.4.2
numpy==2.0.0
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.2
